Add back footer-scripts.js
I accidentally deleted this in a0e8dfde5e520def4b05eb37d17f3722ba52cc27

In my defense, it wasn't at all obvious to me that footer-scripts.js was
not a generated file, but a source file. In future, we should built
files a separate build directory so as not to make mistakes like this
again.

In a future commit, I will modify gulpfile.js to build the output
.min.js and .map files from the source footer-scripts.js file.