Use the_title_attribute instead of the_title for security
Wordpress expects post titles to be HTML, and the_title() prints the
HTML as is, including any <script> if there are any.

Because we will be making posts based on untrusted input, we need to be
careful here. Let's write code that strips out HTML before putting it in
a post title, but let's also use the_title_attribute (which escapes any
HTML), in case we slipped up somewhere and added a post with HTML in its
title. This commit does the second thing.