Added non-DT select
I opted to place a non-DT selection, and to remove the site_link_check elements from a connection in that status. So status of the connection will need to be check outside of the class.

The amount of variable configuration of other systems isn't worth supporting for the system check link. It doesn't affect the integration and use of the class.