Add white and black icons for nav.second-bar and
With a previous commit, an issue was created when the same icon (add-group.svg) was used for a white and then black in different spots in the code. Another icon was needed to be made and placed in the other spot.

The other icons were made to avoid similar issues like this in the future by having the correct icons available.

The naming convention is:
filename.svg = black(or dark) icon with transparent background
filename-2.svg =  white icon with transparent background (for dark backgrounds)
filename-1.svg =  colured icon with transparent background (for dark backgrounds)