issue resolved with a very long testing process with Travis CI. moving forward, whenever we modify sass files, make sure to build, them out and also include the minified .css files in the commit process.
The purpose of this pull request is to make the notification language for past notification dates, more friendly to read.

also, the friendly language as intrernalization applied to it.  With that set in place, hopefully the D.T system can apply appriopriate language adjustments to them as well. Please test.