filtering out duplicates from the people group compact end point.
This occurs when you have a translation setup for a people group, the search is looking for matches in both the title and the translation labels. If your search matches both (ie Title - American and Spanish Label - Americano and you search for American) it will return the people group twice. This compares the post ID to create a unique array that is returned.