autofocus (#1216)
* Added autofocus feature

Users no longer need to manually click on the first box of forms they fill in. This has not been tested extensively, so it may be buggy (but bugs in this case are only as bad as not having the feature at all).

In the unlikely event that a future developer does not want autofocus on a specific element, they may add the `.no-autofocus` class. If they do not want their form to have autofocus at all, they may add the `.no-autofocus` class to the parent element tagged `.reveal`.

Caveats:
* Will not work on HTML elements created after lines 844 - 855 are run
* Has only been tested on a few of the forms, may not work universally

* Update details.js

made the whole deal work at the expense of scalability in the future

* remove console.logs

Co-authored-by: The Intern <>