RTL phone number fix
When the users language is RTL and they enter a phone number with a international prefix with the + the order was not correct. This adds the dir:auto to all inputs and aligns it right for rtl languages. This should allow users to type in RTL or Latin numbers and get things in the proper order and alignment.