Filters menu standardising (#1609)
* Create whatsapp.svg

Adds the missing whatsapp.svg file so it can be used in the appropriate places in a DT instance.

* issue 1366 text size

fix #1366

* Revert "issue 1366 text size"

This reverts commit 6361d97616d69ac88c7eac3cda610ca9662b3b61.

* Standardising menu items and titles

Putting the items in the same order and giving them more clarity. ie. from `all` to `All Groups`

* Use existing translation

* "Add new filter"

Adjusted the phrase "Add new filter" to make it clearer what the link does.